Transition is an interface now
 - Moved `Transition` to `ActionBasedTransition`
 - `Transition` is now an interface
